app/frontend/requirements.txt
streamlit>=1.31.0
requests
aiohttp
pydantic
nanoid
openai
pillow
python-io
fastapi
sqlmodel
